requirements.txt
click==6.7
ipython==5.3.0
ipython-genutils==0.1.0
keras==2.0.2
mako==1.0.6
matplotlib==2.0.0
numpy==1.12.0
pillow==4.0.0
protobuf==3.2.0
pyyaml==3.12
pyxdg==0.25
requests==2.13.0
scipy==0.18.1
sox==1.2.6
tree==0.1.0
theano==0.9.0
 

